Vanier College Page - HTML file
Finished, but missing the nav menu. Not in js, really dont know how make it into js. If you need me to fix that, tell me. The pictures with parallax are being cut off by the "nav menu" that takes 15% width of the screen, tried everything, dont know how to fix that.